fusion-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
python-multipart==0.0.9
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
diffusers==0.26.3
accelerate==0.27.2
onnxruntime==1.17.1
pillow==10.2.0
numpy==1.26.4
prometheus-client==0.20.0
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0
pydantic==2.6.1
python-dotenv==1.0.1
requests==2.31.0
wolframalpha==3.0.1
clip-anytorch==2.6.0
